Move the reference G1 pick policy into molmo_spaces proper
Copies g1_molmo_port/agents/policy_g1ms.py (the reference nav state machine +
GraspPolicy phase machine + WBC control) to
molmo_spaces/policy/solvers/object_manipulation/g1_pick_policy.py, alongside
the other object-manipulation planner policies. Same recipe as the G1Robot
move: the old module becomes a pure re-export shim so every consumer shares a
single class object and a single set of phase constants.

All importers repointed at the new module directly (generate_ported_rollout,
nav_demo, collect_single_main, pick_task_sampler_g1ms); its two intra-port
robot imports now come from molmo_spaces.robots.g1 rather than the
components/robot_g1ms shim.

Also clears two pre-existing ruff findings in pick_task_sampler_g1ms.py that
pre-commit surfaces now that the file is touched (E731 lambda, SIM103).

Gate: byte-identical to the ported-rollout baseline (58/58 lines; steps=2338
sim_time=12.04s success=True), verified after both the copy and the rewire.
